Refactored transposer to separate out logic into a js library.
The logic that transposes text is now contained in transposer-lib.js as a
library that can be used for other purposes. The code was also refactored to be
easier to use (and more elegant). You can now transpose any text by calling
transposeToKey or transposeSemitones depending on your use case. Code specific to
the web app is now contained in transposer.js.